debug

+++ hw_tts/collate_fn/collate.py
@@ -34,11 +34,11 @@ def collate_fn(dataset_items: List[dict]):
         "text": text,
         "mel_len": mel_len,
         "mel_max_len": mel_max_len,
-        "pitch": pad_sequence(pitch, batch_first=True),
-        "energy": pad_sequence(energy, batch_first=True),
-        "mel_pos": pad_sequence(mel_pos, batch_first=True),
-        "src_pos": pad_sequence(src_pos, batch_first=True),
-        "mel_target": pad_sequence(mel, batch_first=True),
-        "alignment": pad_sequence(alignment, batch_first=True),
-        "text_encoded": pad_sequence(text_encoded, batch_first=True)
+        "pitch": pad_sequence(pitch, batch_first=True).long(),
+        "energy": pad_sequence(energy, batch_first=True).long(),
+        "mel_pos": pad_sequence(mel_pos, batch_first=True).long(),
+        "src_pos": pad_sequence(src_pos, batch_first=True).long(),
+        "mel_target": pad_sequence(mel, batch_first=True).float(),
+        "alignment": pad_sequence(alignment, batch_first=True).int(),
+        "text_encoded": pad_sequence(text_encoded, batch_first=True).long()
     }
